brand add

+++ app/Repositories/ProductRepository.php
@@ -4,6 +4,7 @@ namespace App\Repositories;
 
 use App\Interfaces\ProductRepositoryInterface;
 use App\Models\Product;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class ProductRepository implements ProductRepositoryInterface
 {
@@ -14,14 +15,14 @@ class ProductRepository implements ProductRepositoryInterface
         $this->model = $product;
     }
 
-    public function all()
+    public function all($page = 1,$perPage = 10)
     {
-        return $this->model->all();
+        return $this->model->paginate($perPage, ['*'], 'page', $page);
     }
 
     public function find($id)
     {
-        return $this->model->findOrFail($id);
+        return $this->model->find($id);
     }
 
     public function create(array $attributes)
@@ -32,14 +33,22 @@ class ProductRepository implements ProductRepositoryInterface
     public function update($id, array $attributes)
     {
         $product = $this->find($id);
-        $product->update($attributes);
+        if($product != null)
+        {
+            $product->update($attributes);
+            return $product;
+        }
         return $product;
     }
 
     public function delete($id)
     {
         $product = $this->find($id);
-        $product->delete();
-        return $product;
+        if($product != null)
+        {
+            $product->delete();
+            return true;
+        }
+        return false;
     }
 }
